refactor: delete useless object
